news-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
uvloop==0.19.0
httptools==0.6.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
email_validator  # 또는 email-validator==특정버전
passlib[bcrypt]==1.7.4 # 최신 버전 검토 권장
shortuuid==1.0.13
python-jose[cryptography] # 또는 python-jose[cryptography]==특정버전
redis==5.0.4 # 표준 redis-py 사용 시
httpx==0.27.0
wordcloud==1.9.4
matplotlib~=3.8.4 # Python 3.12 호환 버전으로 업데이트
numpy~=1.26.4    # Python 3.12 호환 버전으로 업데이트 (또는 ~=2.0.1)
pandas~=2.2.2     # Python 3.12 호환 버전으로 업데이트
selenium==4.15.0
koNLPy>=0.6.0
JPype1>=1.4.0

